split CLI option handlers
